pyproject.toml
[tool.poetry]
name = "emol"
version = "0.1.0"
description = "Ealdormere's online card system"
authors = ["Ealdormere Code Monkeys"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8.10"
Django = "^4.1.7"
authlib = "^1.2.0"
requests = "^2.28.2"
pytz = "^2022.7.1"
python-dateutil = "^2.8.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
gunicorn = "^20.1.0"
boto3 = "^1.26.79"
celery = "^5.2.7"
mysqlclient = "^2.1.1"
uvicorn = "^0.20.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
bandit = "^1.7.4"
flake8 = "^6.0.0"
pylint = "^2.16.2"

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = "88"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
